Move test to jdk 9
