+++ test/files/jvm/scala-concurrent-tck-b.check
@@ -1,0 +1,2 @@
+starting testUncaughtExceptionReporting
+finished testUncaughtExceptionReporting
